+++ lib/gl_lint/cli.rb
