+++ lib/gl_lint/file_selector.rb
