+++ lib/gl_lint/linter.rb
