+++ lib/gl_lint/version.rb
