+++ spec/gl_lint/file_selector_spec.rb
