+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'g_l_lint'
+require 'gl_lint'
 
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
